Minor changes to update repo for questionnaire.
Automatic pre-commit migration. Also replaced yapf and prospector by
ruff.
Rename statistics.py -> aiida_statistics.py, otherwise, LSP in IDE
complains that stdlib module `statistics` is being overwritten by file.
Removed shebang in favor of executing via `verdi run` (?)
Update author, mail, url in `setup.json`.
query function seems to still be fine for AiiDA v2.5, some null entries
for data types that don't have `process_type` property and
`version.core/plugin` attributes, but I guess that's fine.
