+++ setup.json
@@ -1,36 +1,26 @@
 {
-    "name": "aiida-statistics",
-    "author": "Leopold Talirz",
-    "author_email": "leopold.talirz@epfl.ch",
-    "description": "Compute statistics for AiiDA databases",
-    "url": "https://github.com/ltalirz/aiida-zeopp",
-    "license": "Creative Commons",
-    "classifiers": [
-        "Programming Language :: Python",
-        "Programming Language :: Python :: 2",
-        "Programming Language :: Python :: 3",
-        "Framework :: AiiDA"
-    ],
-    "version": "0.1.0",
-    "install_requires": [
-        "six"
-    ],
-    "extras_require": {
-        "aiida-1.0": [
-            "aiida-core~=1.0.0"
-        ],
-        "aiida-0.12": [
-            "aiida-core~=0.12"
-        ],
-        "pre-commit": [
-            "astroid==1.6.6; python_version<'3.0'",
-            "astroid==2.2.5; python_version>='3.0'",
-            "pre-commit==1.17.0",
-            "yapf==0.28.0",
-            "prospector==1.1.7",
-            "pylint-django==0.11.1; python_version<'3.0'",
-            "pylint==1.9.4; python_version<'3.0'",
-            "pylint==2.3.1; python_version>='3.0'"
-        ]
-    }
+  "name": "aiida-statistics",
+  "author": "aiidateam",
+  "author_email": "developers@aiida.net",
+  "description": "Compute statistics for AiiDA databases",
+  "url": "https://github.com/aiidateam/aiida-core",
+  "license": "Creative Commons",
+  "classifiers": [
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 2",
+    "Programming Language :: Python :: 3",
+    "Framework :: AiiDA"
+  ],
+  "version": "0.2.0",
+  "install_requires": ["six"],
+  "extras_require": {
+    "aiida-2.0": ["aiida-core~=2.0.0"],
+    "aiida-1.0": ["aiida-core~=1.0.0"],
+    "aiida-0.12": ["aiida-core~=0.12"],
+    "pre-commit": [
+      "astroid==1.6.6; python_version<'3.0'",
+      "astroid==2.2.5; python_version>='3.0'",
+      "pre-commit~=2.21.0"
+    ]
+  }
 }
